azure_functions/requirements.txt
azure-functions
requests
openai
azure-data-tables>=12.4.0
